requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
click==8.1.3
Flask==2.1.2
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MyApplication==0.1.0
openapi-schema-pydantic==1.2.4
psycopg2-binary
pydantic==1.9.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
swagger-ui-py==21.12.8
typing_extensions==4.3.0
Werkzeug==2.1.2
flask_cors
python-dotenv
pyTelegramBotAPI
daiquiri

